Update structure to use modules
