+++ src/index.html
@@ -32,12 +32,14 @@
             type="text"
             name="user"
             placeholder="Your name"
+            required
           />
           <input
+            class="form__element form__element--score"
             type="number"
             name="score"
             placeholder="Your score"
-            class="form__element form__element--score"
+            required
           />
           <button class="form__element form__button" type="submit">
             Submit
